# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: PSMS API
  version: 1.0.0
  description: |
    Placement & Service Management System — HTTP API contract.

    This file is the single source of truth for the API surface. Generate
    TypeScript types from this file with `pnpm types:generate` (the result
    is written to `packages/shared/src/types.ts`). Generate the web client
    with `pnpm api:client`.

    Every endpoint declares its required role via the `x-required-role`
    extension. Endpoints with `security: []` are public.

    Requirement traceability uses the IDs defined in `docs/specs/`:
    INP-NN inputs, PRC-NN processing, OUT-NN outputs.
  contact:
    name: Axiom AI Solutions
    email: dorien@axiom.ai
  license:
    name: Proprietary
servers:
  - url: https://psms.knox.edu.jm/api/v1
    description: Production
  - url: https://staging.psms.knox.edu.jm/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local dev
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: psms_session
      description: |
        Session cookie set by `POST /auth/sign-in`. HttpOnly, Secure, SameSite=Lax.
        TTL 8 hours of inactivity. Used for Knox-authenticated roles
        (student, coordinator, admin).

    magicLinkToken:
      type: apiKey
      in: query
      name: token
      description: |
        Single-use 32-byte URL-safe token used by external supervisors.
        Issued by `POST /auth/magic-link/issue`, consumed by any endpoint
        carrying `security: [magicLinkToken: []]`. 24-hour TTL.

  # ───────────────────────────────────────────────────────────────────────
  # Reusable parameters
  # ───────────────────────────────────────────────────────────────────────
  parameters:
    pageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSizeParam:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  responses:
...
    UnprocessableEntity:
      description: Request is well-formed but violates a business rule.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        # PSMS extensions
        request_id: { type: string, description: "ULID, propagated in logs" }
        errors:
          type: array
          description: "Field-level validation errors when status = 400."
          items:
            type: object
            properties:
              path: { type: string, example: "body.email" }
              message: { type: string, example: "Invalid email" }
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    ApplicationStatus:
      type: string
      enum: [SUBMITTED, UNDER_REVIEW, APPROVED, DECLINED, WITHDRAWN]
    Application:
      type: object
      required: [application_id, student_id, opportunity_id, status, submitted_at]
      properties:
        application_id: { type: string, format: uuid }
        student_id: { type: string, format: uuid }
        opportunity_id: { type: string, format: uuid }
        opportunity_title: { type: string, readOnly: true }
        organisation_name: { type: string, readOnly: true }
        motivation: { type: string, maxLength: 2000 }
        cv_url: { type: string, format: uri }
        status: { $ref: '#/components/schemas/ApplicationStatus' }
        submitted_at: { type: string, format: date-time }
        decided_at: { type: string, format: date-time, nullable: true }
        decline_reason: { type: string, nullable: true }
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          readOnly: true
          description: "Matching score 0–1, computed on demand by PRC-02."
    ApplicationInput:
      type: object
      required: [opportunity_id, motivation, cv_url]
      properties:
        opportunity_id: { type: string, format: uuid }
        motivation: { type: string, minLength: 50, maxLength: 2000 }
        cv_url: { type: string, format: uri, description: "Pre-signed upload URL returned by POST /uploads/cv" }
    ApplicationDecision:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVE, DECLINE]
        decline_reason:
          type: string
          maxLength: 1000
          description: "Required when decision = DECLINE."
...
paths:
...
  /applications:
    get:
      tags: [applications]
      summary: List applications visible to the caller
      description: |
        Students see only their own. Coordinators see all. Administrators see all.
      x-required-role: ANY_AUTHENTICATED
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ApplicationStatus' }
        - name: opportunity_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Application' }
    post:
      tags: [applications]
      summary: Submit an application (student)
      x-required-role: STUDENT
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplicationInput' }
      responses:
        '201':
          description: Submitted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Student is not eligible for this opportunity, or has already applied.
  /applications/{application_id}:
    parameters:
      - name: application_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [applications]
      x-required-role: ANY_AUTHENTICATED
      responses:
        '200':
          description: One
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
  /applications/{application_id}/decision:
    post:
      tags: [applications]
      summary: Approve or decline an application (coordinator)
      description: |
        On APPROVE: creates a Placement record, issues a magic-link to the
        supervisor (PRC-03), and dispatches notifications (PRC-07).
      x-required-role: COORDINATOR
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplicationDecision' }
      responses:
        '200':
          description: Decision recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
  /applications/{application_id}/withdraw:
    post:
      tags: [applications]
      summary: Withdraw an application (student-initiated)
      x-required-role: STUDENT
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Withdrawn
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }

  # ── Placements (INP-05, INP-08) ─────────────────────────────────────────
...